genai_api/requirements.txt
fastapi
httpx
openai
google-generativeai
python-dotenv
uvicorn
